backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-google-genai
google-generativeai
pydantic
python-dotenv
pytest
httpx
